+++ frontend/src/App.jsx
@@ -1,12 +1,10 @@
 import React, { Component } from "react";
 import {BrowserRouter, Switch} from 'react-router-dom';
 import {Route} from 'react-router';
-import { DigitHeader } from "@cthit/react-digit-components";
 
 import Home from './home/home';
 import NotFound from './notfound/notfound';
 import Suggestion from './suggestion/Suggestion';
-import Test from "./use-cases/test";
 import "./App.css";
 
 class App extends Component {
